==> wrangle/21_review_groupby_business.ipynb <==
# Review

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import sys
import os
sys.path.append(os.path.abspath(r"C:\ThacSi\HoachDinhCNTT\HDCNTT"))  # Thêm đường dẫn cha của thư mục wrangle
import utils, plot_help

from functools import reduce



%matplotlib inline

#modify read_limit and chunk_size parameters based on memory capacity
df = utils.chunk_loader('../data/review.json', read_limit=-1)

#extract year from datetime series
#make as type string for easier manipulation
df['year'] = df.date.dt.year.astype('str')
df.head()

## Group by business ID + Year and take the mean of reactions to the business

# Chỉ giữ lại các cột số
numeric_cols = ['stars', 'useful', 'funny', 'cool']

# Group theo business_id và year, rồi tính trung bình các cột số
review_year = df.groupby(['business_id', 'year'], as_index=False)[numeric_cols].mean()

# Sắp xếp kết quả nếu muốn
review_year = review_year.sort_values(by=['business_id', 'year'])

# Xem kết quả
review_year.head()


print(df.dtypes)


## It is said that a business is successful if it can survive the test of time. We can use the review data to infer how a business is doing over the years. 
## For example we can measure the average change in star rating over the years. In the abscense of data in the year prior, assume the rating is unchanged since the last time it was recorded.

review_year.pivot(index='business_id', columns='year', values='stars').head()


def feature_year_change(df_year, index, columns, values):
    
    #pivot table to evaluate change in stars
    df_year_feature = df_year.pivot(index=index, 
                                      columns=columns, 
                                      values=values)
    
    #convert to list of lists
    df_feature_list = df_year_feature.values

    df_no_nan = []

    #drop null values
    for array in df_feature_list:
        df_no_nan.append([x for x in array if pd.isna(x) != True])   

    df_no_nan = np.array([np.array(x) for x in df_no_nan])
    
    #save on memory
    del df_feature_list

    #get change in feature
    diffs = []
    for x in df_no_nan:
        if len(x)>2:
            diffs.append(np.diff(x))
        else:
            diffs.append(0)
    
    #again save memory
    del df_no_nan
    
    #get average change in feature value
    feature_change = []
    for x in diffs:
        feature_change.append(np.mean(x))
        
    #return as pandas object
    return pd.DataFrame.from_dict({'index': df_year_feature.index.tolist(),
                                   values: feature_change})


def feature_year_change(df_year, index, columns, values):
    # pivot để tạo bảng doanh nghiệp x năm
    df_year_feature = df_year.pivot(index=index, columns=columns, values=values)

    # danh sách các vector theo từng hàng (doanh nghiệp)
    df_feature_list = df_year_feature.values.tolist()

    # loại bỏ nan
    df_no_nan = [np.array([v for v in x if not pd.isna(v)]) for x in df_feature_list]

    # tính diff (chênh lệch giữa các năm)
    diffs = [np.diff(x) if len(x) > 1 else [0] for x in df_no_nan]

    # tính trung bình thay đổi
    feature_change = [np.mean(x) if len(x) > 0 else 0 for x in diffs]

    return pd.DataFrame.from_dict({
        'index': df_year_feature.index.tolist(),
        'values': feature_change
    })

feat_year_list = []

for feature in ['cool', 'funny', 'stars', 'useful']:
    feat_year_list.append(feature_year_change(review_year, 'business_id','year', feature))

feat_year_list[0].head()

feat_year_list = []

for feature in ['cool', 'funny', 'stars', 'useful']:
    df_feat = feature_year_change(review_year, 'business_id', 'year', feature)
    df_feat.columns = ['index', feature]  # rename cột cho thống nhất
    feat_year_list.append(df_feat)

from functools import reduce

df_year_change = reduce(lambda left, right: pd.merge(left, right, on='index', how='outer'), feat_year_list)

# Đổi tên các cột
df_year_change = df_year_change.rename(columns={
    'index': 'business_id',
    'cool': 'cool_change',
    'funny': 'funny_change',
    'stars': 'stars_change',
    'useful': 'useful_change'
})


df_year_change.head()

df_year_change.to_csv('../data/cleaned/review_year_change.csv')
